Add files via upload
Inhomogeneous Wind Scheme for Urban Street (IWSUS) is developed for genrating wind profiles at typical position of a long street. The IWSUS.m file includes the equations of these wind profiles and am example for a daily variation of heat flux with the use of daily urban surface temperature data in data folder.